@@ -1,10 +1,11 @@
-import React, { useState } from "react";
-import { Route, Switch } from "react-router-dom";
+import { useState } from "react";
+import { Routes, Route } from "react-router-dom";
 import NavBar from "./NavBar";
 import MoviesPage from "./MoviesPage";
+import { MovieList } from "../types";
 
 function App() {
-  const [movies, setMovies] = useState({
+  const [movies, setMovies] = useState<MovieList>({
     1: { id: 1, title: "A River Runs Through It" },
     2: { id: 2, title: "Se7en" },
     3: { id: 3, title: "Inception" },
@@ -13,14 +14,14 @@ function App() {
   return (
     <div>
       <NavBar />
-      <Switch>
+      <Routes>
         <Route path="/movies">
           <MoviesPage movies={movies} />
         </Route>
-        <Route exact path="/">
+        <Route path="/">
           <div>Home</div>
         </Route>
-      </Switch>
+      </Routes>
     </div>
   );
 }
