@@ -1,8 +1,13 @@
 import React from "react";
 import { Route } from "react-router-dom";
 import MoviesList from "./MoviesList";
+import { MovieList } from "../types";
 
-function MoviesPage({ movies }) {
+interface Props {
+  movies: MovieList;
+}
+
+function MoviesPage({ movies }: Props) {
   return (
     <div>
       <MoviesList movies={movies} />
